Windows compatibility fix.
It comes with great amazement that this simple commit makes pyzor
(at least the client) compatible with Windows machines. It is
difficult to fathom why anyone would use pyzor on a Windows machine.
But such a small commit is a small price to pay for the extra
"trove" classifier in the setup script.