Refs. #23. Do reports and whitelists in a single step for MySQL.
Each server engine can now define two methods ``report`` and
``whitelist`` which accepts a list of digests for which the report
and whitelist count should be incremented. If the engines supports
this it must have the ``handles_one_step`` argument set to True.

Implement this for the MySQL engine using INSERT ... ON DUPLICATE KEY
UPDATE, and executemany for batches.

Also add some functional tests around this functionality and fix
the existing unittest.