Wait for the pyzord process to exit before starting running the tests.
This was not an issue before because we used SIGKILL, which terminated
the process immediatly. However now we're using SIGTERM and we're
performing a clean shutdown which in some cases may take a bit longer.

Interestingly PyPy doesn't seem affected for some reason. This is either
because stopping PyPy is faster, or starting PyPy is slower.