Refs. #29. Switch to v1 of redis engine.
Deprecate the old version of string encoding record in favour of
using hashes. Check for deprecated records when the engine is
created and display warnings for the user.

Add ability to delete old records to the pyzor-migrate script, in
order to get rid of the warning.

Use hincrby when incrementing reports and whitelists counts, instead
of replacing the whole record. Make sure to set propper expiry if
necessary.

Old version will still be available under the `redis_v0` name.