scripts/pyzor: read stdin as binary in _get_input_msg().
Reading stdin in python-3.x is done as text, with a best-guess
encoding. But this can go awry: for example, if an iso-8859-1 message
is passed in and if python guesses the "utf-8" encoding, then read()
will fail with a UnicodeDecodeError on non-ASCII characters. For
example, the "copyright" symbol is a single byte 0xa9 in iso-8859-1,
and the utf-8 decoder can't handle it:

  UnicodeDecodeError: 'utf-8' codec can't decode byte 0xa9... invalid
  start byte

Instead -- and as was done in python-2.x -- we can read stdin as
binary using the new get_binary_stdin() function. Afterwards, we use
email.message_from_bytes() instead of the email.message_from_file()
constructor to parse the byte data. The resulting function is able to
correctly parse these messages.

Closes: https://github.com/SpamExperts/pyzor/issues/64